requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-yasg==1.21.8
environs==14.1.0
gunicorn==23.0.0
inflection==0.5.1
marshmallow==3.25.1
packaging==24.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.8.2
